requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
multidict==6.0.5
openai==1.12.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
yarl==1.9.4
pip~=23.3.2
wheel~=0.42.0
filelock~=3.13.1
PyYAML~=6.0.1
claudetools~=0.5.0
setuptools~=69.0.2
fsspec~=2024.3.0
six~=1.16.0
python-dateutil~=2.8.2
tokenizers~=0.15.2
anthropic~=0.23.1